requirements.txt
# requirements.txt - generated file
streamlit
pymysql
Pillow
